refactor(PKGenerationType): annotaion값 추출 로직 enum으로 이동
