--- src/main/java/persistence/sql/ddl/PKGenerationType.java
@@ -1,25 +1,0 @@
-package persistence.sql.ddl;
-
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-
-public enum PKGenerationType {
-    IDENTITY("generated by default as identity");
-
-    private final String sql;
-
-    PKGenerationType(final String sql) {
-        this.sql = sql;
-    }
-
-    public static PKGenerationType get(final GeneratedValue generatedValue) {
-        if (generatedValue.strategy().equals(GenerationType.IDENTITY)) {
-            return IDENTITY;
-        }
-        throw new IllegalArgumentException("Unsupported GenerationType");
-    }
-
-    public String getSql() {
-        return sql;
-    }
-}
